requirements.txt
streamlit==1.42.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
matplotlib==3.9.4
seaborn==0.13.2
wordcloud==1.9.4
PySastrawi==1.2.0
google-play-scraper==1.2.7
nltk==3.9.1
